requirements.txt
streamlit
pandas
numpy
joblib
plotly
tensorflow
scikit-learn
python-dotenv
groq
